fix(app-shell): clear frontend lint blockers for workflow Diff work
Drop the leftover reviewCloseRequestId reset, sync refs outside render,
seed deploy dialog name without setState-in-effect, and extract plugin
search filtering so react-refresh stays happy.
